# Conflicts:
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/com/android_group10/needy/MainActivity.java

+++ app/src/main/java/com/android_group10/needy/MainActivity.java
@@ -1,19 +1,12 @@
 package com.android_group10.needy;
 
 import android.annotation.SuppressLint;
-import android.app.Activity;
-import android.content.ActivityNotFoundException;
 import android.content.Intent;
-
-import android.database.sqlite.SQLiteDatabase;
 import android.graphics.Bitmap;
 import android.graphics.Color;
 import android.graphics.drawable.ColorDrawable;
-
-
 import android.net.Uri;
 import android.os.Bundle;
-
 import android.os.Environment;
 import android.provider.MediaStore;
 import android.util.Log;
@@ -21,45 +14,34 @@ import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
-import android.widget.EditText;
 import android.widget.ImageView;
 import android.widget.PopupMenu;
 import android.widget.TextView;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
-import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.annotation.Nullable;
 import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 import androidx.core.content.FileProvider;
 import androidx.core.view.GravityCompat;
 import androidx.drawerlayout.widget.DrawerLayout;
-import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentTransaction;
 import androidx.navigation.NavController;
-import androidx.navigation.Navigation;
+import androidx.navigation.NavDestination;
+import androidx.navigation.fragment.NavHostFragment;
 import androidx.navigation.ui.AppBarConfiguration;
 import androidx.navigation.ui.NavigationUI;
 
-
 import com.android_group10.needy.LocalDatabase.DbBitmapUtility;
 import com.android_group10.needy.LocalDatabase.LocalDatabaseHelper;
-import com.android_group10.needy.ui.InNeed.AddPostRecordFragment;
-import com.android_group10.needy.ui.InNeed.InNeedFragment;
+import com.android_group10.needy.ui.InNeed.InNeedFragmentDirections;
 import com.android_group10.needy.ui.LogInAndRegistration.LogIn;
-import com.android_group10.needy.ui.NeedsAndDeeds.NeedsAndDeedsFragment;
-
-import com.android_group10.needy.ui.Profile.ProfileFragment;
-import com.android_group10.needy.ui.messaging.MessagingFragment;
 import com.bumptech.glide.Glide;
 import com.facebook.CallbackManager;
 import com.facebook.FacebookSdk;
 import com.facebook.login.LoginManager;
-import com.facebook.share.model.ShareHashtag;
 import com.facebook.share.model.ShareLinkContent;
-import com.facebook.share.widget.ShareButton;
 import com.facebook.share.widget.ShareDialog;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.android.material.navigation.NavigationView;
@@ -71,6 +53,7 @@ import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
+import com.rom4ek.arcnavigationview.ArcNavigationView;
 
 import java.io.File;
 import java.io.IOException;
@@ -80,92 +63,64 @@ import java.util.Date;
 import java.util.Locale;
 
 
-public class MainActivity extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener, PopupMenu.OnMenuItemClickListener {
+public class MainActivity extends AppCompatActivity implements PopupMenu.OnMenuItemClickListener, NavController.OnDestinationChangedListener {
 
-
-    private FragmentTransaction fragmentTransaction;
-    private FragmentManager fragmentManager;
-    private AppBarConfiguration mAppBarConfiguration;
-    private DrawerLayout drawer;
-    private Toolbar toolbar;
-    private NavigationView navigationView;
-    private FloatingActionButton floatingActionButton;
-    private FirebaseAuth mAuth;
-    private DatabaseReference reference;
+    private DrawerLayout drawerLayout;
+    AppBarConfiguration appBarConfiguration;
+    private NavController navController;
+    FloatingActionButton floatingActionButton;
     private String userId;
-    private FirebaseUser user;
-    private FirebaseDatabase db = FirebaseDatabase.getInstance();
-    private FirebaseAuth firebaseAuth;
-    private EditText phoneNumber_dialog, city_dialog, zipCode_dialog;
-    private View header;
+    private View headerView;
     private ImageView profileImage;
-    private boolean hasChanged = false;
-    private Bitmap bitmap;
     private long pressedTime;
 
     private String imageFilePath;
     File photoFile = null;
-    private int five = 5;
-    private SQLiteDatabase database;
     private Uri selectedImageUri;
     private LocalDatabaseHelper localDatabaseHelper;
     private ProfilePictureManager ppManager;
 
     private static final int SELECT_PICTURE = 100;
     private static final int CAMERA_REQUEST = 1;
-    private static final int SHARE_ON_FACEBOOK = 2;
+    //private static final int SHARE_ON_FACEBOOK = 2;
 
     private static final String TAG = "StoreImageActivity";
-    private Activity thisActivity = this;
     ShareDialog shareDialog;
     CallbackManager callbackManager;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
-        localDatabaseHelper = new LocalDatabaseHelper(this);
-        ppManager = new ProfilePictureManager();
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+        localDatabaseHelper = new LocalDatabaseHelper(this);
+        ppManager = new ProfilePictureManager();
         initialization();
         updateHeader();
         facebookSDKInitialize();
-        setSupportActionBar(toolbar);
-        floatingActionButton = findViewById(R.id.fab);
-        floatingActionButton.setOnClickListener(view -> Snackbar.make(view, "Click here to add a help request", Snackbar.LENGTH_LONG)
-                .setAction("New Need", new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        fragmentManager = getSupportFragmentManager();
-                        fragmentTransaction = fragmentManager.beginTransaction();
-                        fragmentTransaction.replace(R.id.fragment_container, new AddPostRecordFragment());
-                        fragmentTransaction.addToBackStack(null);
-                        fragmentTransaction.commit();
-                    }
-                }).show());
-        floatingActionButton.setImageResource(R.drawable.add_icon);
 
-        ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(this, drawer, toolbar
+        //Don't need this, handled automatically
+        /*ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(this, drawer, toolbar
                 , R.string.navigation_drawer_open, R.string.navigation_drawer_close);
         drawer.addDrawerListener(toggle);
         toggle.setDrawerIndicatorEnabled(true);
-        toggle.syncState();
+        toggle.syncState();*/
 
-        navigationView.setNavigationItemSelectedListener(this);
+        //navigationView.setNavigationItemSelectedListener(this);
 
-        fragmentManager = getSupportFragmentManager();
+        /*fragmentManager = getSupportFragmentManager();
         fragmentTransaction = fragmentManager.beginTransaction();
-        fragmentTransaction.replace(R.id.fragment_container, new InNeedFragment());
-        fragmentTransaction.commit();
+        fragmentTransaction.replace(R.id.nav_host_fragment, new InNeedFragment());
+        fragmentTransaction.commit();*/
 
         profileImage.setOnClickListener(v -> {
-            drawer.closeDrawer(GravityCompat.START);
+            //drawerLayout.closeDrawer(GravityCompat.START); //What is this meant to do?
             showPopup(profileImage);
-            drawer.openDrawer(GravityCompat.START);
+            //drawerLayout.openDrawer(GravityCompat.START); //-^
         });
 
         //this.deleteDatabase("Images.db");
 
-        drawer.addDrawerListener(new DrawerLayout.DrawerListener() {
+        drawerLayout.addDrawerListener(new DrawerLayout.DrawerListener() {
             @Override
             public void onDrawerSlide(@NonNull View drawerView, float slideOffset) {
             }
@@ -173,7 +128,7 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
             @Override
             public void onDrawerOpened(@NonNull View drawerView) {
                 ProfilePictureManager ppManager = new ProfilePictureManager();
-                ppManager.displayProfilePic(thisActivity, profileImage, true);
+                ppManager.displayProfilePic(MainActivity.this, profileImage, true);
             }
 
             @Override
@@ -196,31 +151,38 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
 
     @Override
     public boolean onSupportNavigateUp() {
-        NavController navController = Navigation.findNavController(this, R.id.fragment_container);
-        return NavigationUI.navigateUp(navController, mAppBarConfiguration)
-                || super.onSupportNavigateUp();
+        return NavigationUI.navigateUp(navController, appBarConfiguration) || super.onSupportNavigateUp();
     }
 
-    @SuppressLint("NonConstantResourceId")
+    @Override
+    public void onDestinationChanged(@NonNull NavController controller, @NonNull NavDestination destination, @Nullable Bundle arguments) {
+        if (destination.getId() == R.id.in_need) {
+            floatingActionButton.show();
+        } else {
+            floatingActionButton.hide();
+        }
+    }
+
+    /*@SuppressLint("NonConstantResourceId")
     @Override
     public boolean onNavigationItemSelected(@NonNull MenuItem item) {
 
         switch (item.getItemId()) {
-            case R.id.nav_in_need:
+            case R.id.in_need:
                 changeFragment(new InNeedFragment());
 
                 floatingActionButton.show();
 
                 break;
-            case R.id.nav_needs_and_deeds:
+            case R.id.needs_and_deeds:
                 changeFragment(new NeedsAndDeedsFragment());
                 floatingActionButton.hide();
                 break;
-            case R.id.nav_conversations:
+            case R.id.messaging:
                 changeFragment(new MessagingFragment());
                 floatingActionButton.hide();
                 break;
-            case R.id.nav_profile:
+            case R.id.profile:
                 changeFragment(new ProfileFragment());
                 floatingActionButton.hide();
                 break;
@@ -237,83 +199,132 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
                 break;
         }
         //This helps to close the navigation bar after choose an item from it.
-        drawer.closeDrawer(GravityCompat.START);
+        drawerLayout.closeDrawer(GravityCompat.START);
         return true;
-    }
+    }*/
 
-    // Change fragments when press on navigation drawer.
+    /*// Change fragments when press on navigation drawer.
     public void changeFragment(Fragment fragment) {
         fragmentManager = getSupportFragmentManager();
         fragmentTransaction = fragmentManager.beginTransaction();
-        fragmentTransaction.replace(R.id.fragment_container, fragment);
+        fragmentTransaction.replace(R.id.nav_host_fragment, fragment);
         fragmentTransaction.commit();
-    }
+    }*/
 
     // Item menu
     @Override
     public boolean onOptionsItemSelected(@NonNull MenuItem item) {
         if (item.getItemId() == R.id.about_us) {
             about_us_dialog();
+            return true;
         }
 
         return super.onOptionsItemSelected(item);
     }
 
-
     // Can press back button and only close the navigation bar and not the activity
     @Override
     public void onBackPressed() {
-        if (drawer.isDrawerOpen(GravityCompat.START)) {
-            drawer.closeDrawer(GravityCompat.START);
+        //This breaks the navigation
+        /*if (drawerLayout.isDrawerOpen(GravityCompat.START)) {
+            drawerLayout.closeDrawer(GravityCompat.START);
         } else if (pressedTime + 2000 > System.currentTimeMillis()) {
             super.onBackPressed();
             finish();
         } else {
             Toast.makeText(getBaseContext(), "Press back again to exit", Toast.LENGTH_SHORT).show();
             pressedTime = System.currentTimeMillis();
+        }*/
+
+        //This way the back stack behaves normally, TODO: find a way to also implement double tap exit
+        if (drawerLayout.isDrawerOpen(GravityCompat.START)){
+            drawerLayout.closeDrawer(GravityCompat.START);
+        } else {
+            super.onBackPressed();
         }
 
     }
 
     public void initialization() {
-        mAuth = FirebaseAuth.getInstance();
-        drawer = findViewById(R.id.drawer_layout);
-        navigationView = findViewById(R.id.nav_view);
-        toolbar = findViewById(R.id.toolbar);
+        drawerLayout = findViewById(R.id.drawer_layout);
+        ArcNavigationView navigationView = findViewById(R.id.nav_view);
 
+        //Set-up appbar
+        Toolbar toolbar = findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+        appBarConfiguration =
+                new AppBarConfiguration.Builder(R.id.in_need, R.id.needs_and_deeds, R.id.messaging, R.id.profile)
+                        .setOpenableLayout(drawerLayout)
+                        .build();
 
-        header = navigationView.getHeaderView(0);
-        profileImage = header.findViewById(R.id.profileImageHeader);
+        //Initialize navigation
+        NavHostFragment navHostFragment = (NavHostFragment) getSupportFragmentManager().findFragmentById(R.id.nav_host_fragment);
+        if (navHostFragment != null) {
+            navController = navHostFragment.getNavController();
+            NavigationUI.setupActionBarWithNavController(this, navController, appBarConfiguration);
+            NavigationUI.setupWithNavController(navigationView, navController);
+        }
 
-        phoneNumber_dialog = findViewById(R.id.Phone_dialog);
-        city_dialog = findViewById(R.id.City_dialog);
-        zipCode_dialog = findViewById(R.id.zip_dialog);
+        //Initialize logout and share buttons in drawer
+        //Logout
+        navigationView.getMenu().findItem(R.id.log_out).setOnMenuItemClickListener(menuItem -> {
+            FirebaseAuth.getInstance().signOut();
+            LoginManager.getInstance().logOut();
+            startActivity(new Intent(this, LogIn.class));
+            finish();
+            return true;
+        });
+        //Share
+        navigationView.getMenu().findItem(R.id.share).setOnMenuItemClickListener(menuItem -> {
+            if (ShareDialog.canShow(ShareLinkContent.class)) {
+                ShareOnFacebook shareOnFacebook = new ShareOnFacebook();
+                shareDialog.show(shareOnFacebook.createShare());
+            }
+            return true;
+        });
+
+        //Initialize fab
+        floatingActionButton = findViewById(R.id.fab);
+        floatingActionButton.setOnClickListener(view -> Snackbar.make(view, "Click here to add a help request", Snackbar.LENGTH_LONG)
+                .setAction("New Need", v -> {
+                    navController.navigate(InNeedFragmentDirections.actionInNeedToAddPost());
+
+                    //Old transaction
+                    /*fragmentManager = getSupportFragmentManager();
+                    fragmentTransaction = fragmentManager.beginTransaction();
+                    fragmentTransaction.replace(R.id.nav_host_fragment, new AddPostRecordFragment());
+                    fragmentTransaction.addToBackStack(null);
+                    fragmentTransaction.commit();*/
+                }).show());
+        floatingActionButton.setImageResource(R.drawable.add_icon);
+        //Set listener for fab behaviour based on fragment
+        navController.addOnDestinationChangedListener(this);
+
+        headerView = navigationView.getHeaderView(0);
+        profileImage = headerView.findViewById(R.id.profileImageHeader);
 
     }
 
 
     //Update the User name and User email in the Header.
     public void updateHeader() {
-        NavigationView navigationView = findViewById(R.id.nav_view);
-        View headerView = navigationView.getHeaderView(0);
         TextView userNameOnHeader = headerView.findViewById(R.id.user_name);
         TextView userEmailOnHeader = headerView.findViewById(R.id.user_email_header);
 
-        user = FirebaseAuth.getInstance().getCurrentUser();
-        reference = FirebaseDatabase.getInstance().getReference("Users");
-        userId = user.getUid();
+        FirebaseUser user = FirebaseAuth.getInstance().getCurrentUser();
+        DatabaseReference reference = FirebaseDatabase.getInstance().getReference("Users");
+        userId = user != null ? user.getUid() : "ERROR_UID_MISSING";
 
         reference.child(userId).addListenerForSingleValueEvent(new ValueEventListener() {
-            @SuppressLint("SetTextI18n")
             @Override
             public void onDataChange(@NonNull DataSnapshot snapshot) {
                 User user = snapshot.getValue(User.class);
 
                 if (user != null) {
-                    String name = user.getFirstName();
+                    String name = "Welcome " + user.getFirstName();
                     String email = user.getEmail();
 
-                    userNameOnHeader.setText("Welcome " + name);
+                    userNameOnHeader.setText(name);
                     userEmailOnHeader.setText(email);
                     //loadImageFromDB();
                 }
@@ -334,11 +345,6 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
         alertD.getWindow().setBackgroundDrawable(new ColorDrawable(Color.TRANSPARENT));
         alertD.setView(view);
         alertD.show();
-    }
-
-    @Override
-    protected void onStart() {
-        super.onStart();
     }
 
     // Header Photo methods.
@@ -380,29 +386,22 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
 
     }
 
-    @SuppressLint("NonConstantResourceId")
     @Override
     public boolean onMenuItemClick(MenuItem item) {
 
-        switch (item.getItemId()) {
-            case R.id.change_image:
-
-                //This will open the gallery
-                openImageChooser();
-                return true;
-
-            case R.id.delete_image:
-                localDatabaseHelper.deleteImage(userId);
-                profileImage.setImageResource(R.drawable.anonymous_mask);
-                return true;
-
-            case R.id.take_image:
-                openCamera();
-                return true;
-
-            default:
-                return false;
+        int itemId = item.getItemId();
+        if (itemId == R.id.change_image) {//This will open the gallery
+            openImageChooser();
+            return true;
+        } else if (itemId == R.id.delete_image) {
+            localDatabaseHelper.deleteImage(userId);
+            profileImage.setImageResource(R.drawable.anonymous_mask);
+            return true;
+        } else if (itemId == R.id.take_image) {
+            openCamera();
+            return true;
         }
+        return false;
     }
 
 
@@ -498,12 +497,8 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
     }
 
     protected void facebookSDKInitialize() {
-        FacebookSdk.sdkInitialize(getApplicationContext());
+        //FacebookSdk.sdkInitialize(getApplicationContext()); //Deprecated, done automatically now
         callbackManager = CallbackManager.Factory.create();
         shareDialog = new ShareDialog(this);
     }
-
-    public void preventClicks(View view) {
-    }
-
 }
